Add finance_calculator.py to calculate compound interest

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # User Input for financial Details
-monthly_income = int(input("Enter your monthly income: "))
-Total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+Total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = Total_monthly_expenses - monthly_income
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 print(f"Your monthly savings are ${monthly_savings:.2f}.")
